requirements.txt
Scrapy==2.13.4
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
streamlit==1.52.1
requests==2.32.5
lxml==6.0.2
altair==6.0.0
pydeck==0.9.1
matplotlib-inline==0.2.1
